ultime modifiche da cluster

+++ Libft/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: volivier <volivier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/16 12:54:49 by volivier          #+#    #+#             */
-/*   Updated: 2024/01/18 16:58:58 by volivier         ###   ########.fr       */
+/*   Updated: 2024/01/20 19:36:07 by volivier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
+	size_t			i;
 	char			*ret;
-
 
 	ret = malloc(sizeof(char) * (ft_strlen((char *)s) + 1));
 	if (!s || !ret || !f)
